Разделит код на модули
